Updated BlockChain Project files.
Blockchain Project Version 1.1

When testing the project using Jest, comment out code from line 34 to line 36 in the server.js file or else Jest testing will not work. This is because there will be two instances of the app listening to PORT 3001. 

To run the project after testing, simply uncomment the code.